feat: add execution cache using StagedSubstateStoreManager

+++ core-rust/state-manager/src/transaction/validator_transaction.rs
@@ -60,7 +60,7 @@ impl ValidatorTransaction {
                 },
                 args: scrypto_encode(&ClockSetCurrentTimeInvocation {
                     receiver: CLOCK,
-                    current_time_ms: *timestamp_ms,
+                    current_time_ms: *timestamp_ms as i64,
                 })
                 .unwrap(),
             },
